[irods/irods#4740] Add iadmin modrepl and ls
iadmin modrepl can be used to modify a number of different columns in
R_DATA_MAIN. Some have been restricted until such a time as a valid use
case justifies its exposure in the client. Until then, only a subset of
the columns are modifiable via iadmin modrepl.

iadmin ls now lists the GenQuery attributes and their values which map to
the columns of R_DATA_MAIN. This is mainly for convenience.