[irods/irods#6624] iadmin: Replace SimpleQuery
iadmin has historically used SimpleQuery to fetch information from the
catalog and display it to the user. This fixes an issue with SimpleQuery
which can cause the servicing agent to crash by reimplementing all uses
of SimpleQuery to use GenQuery.

The following operations have been reworked to use GenQuery:
 - lq
 - lr
 - lu
 - lua
 - luan
 - luz
 - lz

This also removes the last reference to rcSimpleQuery in the iCommands.
Because SimpleQuery is only allowed to be used by rodsadmins, the
enforcement of running iadmin commands only by rodsadmins has been moved
to an earlier place in the sequence. Obviously, rodsusers can still run
the queries using iquest, but this has always been the case.