[irods/irods 6879] iput: Disconnect in signal handler
This adds a signal handler to iput in order to allow the client to
disconnect when it receives SIGINT (i.e. ^C). This involves the RcComm
used by the client being turned into a global so that the signal handler
is able to access it. Disconnecting before terminating allows the
servicing agent to stale the in-flight replica immediately rather than
waiting for the connection to time out.

One will still observe some noise on the server side because the most
common way that this fails is by interrupting a parallel transfer and
the threads writing data will disconnect suddenly when the iput process
exits on the client side.