[irods/irods 2978] iadmin: Add support to "add" a "group"
iadmin mkgroup was just using the ("add", "user") form
of the GeneralAdmin API. The API now supports ("add", "group")
for creating groups in order to separate groups and users.
This commit updates iadmin mkgroup such that it uses the
("add", "group") form of the API for making groups if it's
supported.